dom/indexedDB/test/test_file_os_delete.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
  function testSteps()
  {
    const READ_WRITE = IDBTransaction.READ_WRITE;

    const name = window.location.pathname;
    const description = "My Test Database";

    const objectStoreName = "Blobs";

    getUsage(grabFileUsageAndContinueHandler);
    let startUsage = yield;

    const fileData = { key: 1, file: getRandomFile("random.bin", 100000) };

    {
      let request = indexedDB.open(name, 1, description);
      request.onerror = errorHandler;
      request.onupgradeneeded = grabEventAndContinueHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield;

      is(event.type, "upgradeneeded", "Got correct event type");

      let db = event.target.result;
      db.onerror = errorHandler;

      let objectStore = db.createObjectStore(objectStoreName, { });

      objectStore.add(fileData.file, fileData.key);

      event = yield;

      is(event.type, "success", "Got correct event type");

      getUsage(grabFileUsageAndContinueHandler);
      let usage = yield;

      is(usage, startUsage + fileData.file.size, "Correct file usage");

      let trans = db.transaction([objectStoreName], READ_WRITE);
      trans.objectStore(objectStoreName).delete(fileData.key);
      trans.oncomplete = grabEventAndContinueHandler;
      event = yield;

      is(event.type, "complete", "Got correct event type");

      getUsage(grabFileUsageAndContinueHandler);
      usage = yield;

      is(usage, startUsage + fileData.file.size, "OS file exists");

      fileData.file = null;
    }

    scheduleGC();
    yield;

    getUsage(grabFileUsageAndContinueHandler);
    let endUsage = yield;

    is(endUsage, startUsage, "OS file deleted");

    finishTest();
    yield;
  }
  </script>
  <script type="text/javascript;version=1.7" src="file.js"></script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
